Revert "Drop logic updated"
This reverts commit 2c45e588504ac923fdd8f685a018fafd1a72c38a.

+++ Game/Tetrominoe.java
@@ -3,7 +3,6 @@ package Game;
 import java.awt.Color;
 import java.awt.Graphics;
 import java.awt.Graphics2D;
-import java.util.Arrays;
 
 public abstract class Tetrominoe {
 
@@ -56,32 +55,11 @@ public abstract class Tetrominoe {
     }
 
     public void rotateLeft() {
-        if (canRotate) {
 
-        }
     }
 
     public void rotateRight() {
-        if (canRotate) {
-            boolean[][] previousMatrix = new boolean[matrix.length][matrix[0].length];
 
-            for (int i = 0; i < matrix.length; i++) {
-                for (int j = 0; j < matrix[i].length; j++) {
-                    previousMatrix[i][j] = matrix[i][j];
-                }
-            }
-
-            for (int i = 0; i < matrix.length / 2; i++) {
-                for (int j = i; j < matrix.length - i - 1; j++) {
-                    boolean temp = matrix[i][j];
-
-                    matrix[i][j] = matrix[j][matrix.length - i - 1];
-                    matrix[j][matrix.length - i - 1] = matrix[matrix.length - i - 1][matrix.length - j - 1];
-                    matrix[matrix.length - i - 1][matrix.length - j - 1] = matrix[matrix.length - j - 1][i];
-                    matrix[matrix.length - j - 1][i] = temp;
-                }
-            }
-        }
     }
 
     private boolean kick(int xPos) {
@@ -105,6 +83,7 @@ public abstract class Tetrominoe {
 
     public void convert(boolean[][] boardMatrix, Color[][] colorMatrix) {
         int row;
+        boolean rowDeleted = false;
 
         for (int i = 0; i < matrix.length; i++) {
             for (int j = 0; j < matrix[i].length; j++) {
@@ -117,7 +96,7 @@ public abstract class Tetrominoe {
             }
         }
 
-        for (int i = matrix[0].length; i >= 0; i--) {
+        for (int i = 0; i < matrix[0].length; i++) {
             row = 0;
 
             if (i + y < board.getGridY() && i + y >= 0) {
@@ -132,6 +111,8 @@ public abstract class Tetrominoe {
 
                 if (row == board.getGridX()) {
                     System.out.println("delete");
+                    rowDeleted = true;
+
                     for (int k = 0; k < boardMatrix.length; k++) {
                         boardMatrix[k][i + y] = false;
                         colorMatrix[k][i + y] = null;
@@ -140,7 +121,9 @@ public abstract class Tetrominoe {
             }
         }
 
-        board.drop();
+        if (rowDeleted) {
+            board.drop();
+        }
     }
 
     public void draw(Graphics g) {
